+++ src/util/client.rs
@@ -47,25 +47,17 @@ pub async fn get_token_ensure_refreshed(
     user_id: String,
     token: &Token,
     ctx: AppContext,
-) -> crate::Result<AuthCodeSpotify> {
+) -> crate::Result<(AuthCodeSpotify, Token)> {
+    let mut token = token.clone();
     let mut client = create_from_token(token.clone());
 
-    let is_expired = client
-        .get_token()
-        .lock()
-        .await
-        .unwrap()
-        .as_ref()
-        .map(|token| token.is_expired())
-        .is_some_and(|val| val);
-
-    if is_expired {
+    if token.is_expired() {
         // Create new client with our credentials and add our current token
         client = create_oauth_client();
         *client.token.lock().await.unwrap() = Some(token.clone());
 
         // Request a new token
-        let token = client
+        token = client
             .refetch_token()
             .await?
             .ok_or_else(|| anyhow::anyhow!("could not refetch token"))?;
@@ -79,5 +71,5 @@ pub async fn get_token_ensure_refreshed(
     }
 
     // If we requested a new token, the client now has it
-    Ok(client)
+    Ok((client, token))
 }