+++ src/web/middleware/auth.rs
@@ -2,11 +2,10 @@ use crate::{
     context::AppContext,
     repo::user::UserRepo,
     util::{client, jwt},
-    web::router::JWT_COOKIE,
+    web::{router::JWT_COOKIE, session},
     CONFIG,
 };
 use axum::{extract::State, http::Request, middleware::Next, response::IntoResponse};
-use rspotify::AuthCodeSpotify;
 use tower_cookies::Cookies;
 
 pub async fn middleware<B>(
@@ -19,19 +18,29 @@ pub async fn middleware<B>(
         .get(JWT_COOKIE)
         .ok_or_else(|| crate::error::Error::UnauthorizedError)?;
 
-    let client = try_create_auth_client(jwt_cookie.value(), ctx)
+    let session = try_create_auth_session(jwt_cookie.value(), ctx)
         .await
         .map_err(|_| crate::error::Error::UnauthorizedError)?;
 
-    // Add Spotify OAuth client as extension to be accessed by any route that wishes to perform action on user's behalf
-    req.extensions_mut().insert(client);
+    req.extensions_mut().insert(session);
 
     Ok(next.run(req).await)
 }
 
-async fn try_create_auth_client(jwt: &str, ctx: AppContext) -> crate::Result<AuthCodeSpotify> {
+async fn try_create_auth_session(jwt: &str, ctx: AppContext) -> crate::Result<session::Session> {
     let user_id = jwt::verify_jwt(CONFIG.web.jwt_secret.as_ref(), jwt)?;
-    let token = UserRepo::new(ctx.clone()).get_token_by_user_id(&user_id)?;
+    let token_str = &UserRepo::new(ctx.clone()).get_token_by_user_id(&user_id)?;
 
-    client::get_token_ensure_refreshed(user_id, &serde_json::from_str(&token)?, ctx).await
+    let (client, token) = client::get_token_ensure_refreshed(
+        user_id.clone(),
+        &serde_json::from_str(&token_str)?,
+        ctx,
+    )
+    .await?;
+
+    Ok(session::Session {
+        user_id,
+        token,
+        client,
+    })
 }